Fix circular import problem
